Some moves and orgs from fixing up API
